Move Scheme, Host and Global into the server module
